Locale experiment
This patch allows locale keys passed to the main Uppy instance to
override default keys in plugins. It's just one idea to kick things off,
I'm not attached to this particular implementation.

If we do this we need to make sure that all keys have unique names and
that things specific to eg. a single service are "namespaced"
(`transloaditEncoding` instead of `encoding`).

The benefit to this is that there is a central point of configuration
for languages, so there could be language packs with strings for eg
Czech at `@uppy/lang-cz` (or `@uppy/langs/cz`) that would be very easy
to use.